refactor: split domain and time utilities
