+++ scripts/build-static-site.mjs
@@ -1,4 +1,11 @@
-import { cpSync, mkdirSync, readFileSync, rmSync, writeFileSync } from "node:fs";
+import {
+  cpSync,
+  mkdirSync,
+  readFileSync,
+  readdirSync,
+  rmSync,
+  writeFileSync,
+} from "node:fs";
 import { dirname, join } from "node:path";
 import { fileURLToPath } from "node:url";
 
@@ -15,7 +22,11 @@ const indexHtml = readFileSync(join(rootDir, "frontend", "index.html"), "utf8")
 writeFileSync(join(publicDir, "index.html"), indexHtml);
 cpSync(join(rootDir, "frontend", "styles.css"), join(publicDir, "styles.css"));
 cpSync(join(rootDir, "dist", "frontend", "app.js"), join(publicDir, "frontend", "app.js"));
-cpSync(
-  join(rootDir, "dist", "src", "orderController.js"),
-  join(publicDir, "src", "orderController.js")
-);
+
+for (const fileName of readdirSync(join(rootDir, "dist", "src"))) {
+  if (!fileName.endsWith(".js")) {
+    continue;
+  }
+
+  cpSync(join(rootDir, "dist", "src", fileName), join(publicDir, "src", fileName));
+}
